image_metadata/requirements.txt
Pillow
termcolor
argparse